+++ interfaces/PA-App/src/app/personal/personal.page.ts
@@ -38,7 +38,7 @@ import { InclusionStatusComponent } from './../personal-components/inclusion-sta
   styleUrls: ['personal.page.scss'],
 })
 export class PersonalPage implements OnInit {
-  @ViewChild(IonContent, { static: true })
+  @ViewChild(IonContent, { static: false })
   public ionContent: IonContent;
 
   @ViewChild('conversationContainer', { read: ViewContainerRef, static: true })