+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - 2021-09-15: Upgrade PA-App to Angular v8
 - 2021-09-15: Upgrade HO-Portal to Angular v8
+- 2021-09-15: Upgrade AW-App to Angular v8
 
 ---